requirements.txt
asgiref
autopep8
Django
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
gunicorn
whitenoise
pymysql
openpyxl
apache-airflow